requirements.txt
alive-progress == 3.1.5
pillow == 10.2.0
numpy == 1.24.4